+++ CNFGCocoaNSImageDriver.m
@@ -180,7 +180,7 @@ void CNFGHandleInput()
     [app_currDate release];
 }
 
-void CNFGUpdateScreenWithBitmap( unsigned long * data, int w, int h )
+void CNFGUpdateScreenWithBitmap( uint32_t * data, int w, int h )
 {
     [app_pool release];
     app_pool = [NSAutoreleasePool new];
