+++ CNFGWinDriver.c
@@ -58,7 +58,7 @@ void CNFGGetDimensions( short * x, short * y )
 }
 
 
-void CNFGUpdateScreenWithBitmap( unsigned long * data, int w, int h )
+void CNFGUpdateScreenWithBitmap( uint32_t * data, int w, int h )
 {
 	RECT r;
